Deduplicate StripProto methods
