+++ src/google/protobuf/compiler/java/java_name_resolver.cc
@@ -33,6 +33,7 @@
 #include <map>
 #include <string>
 
+#include <google/protobuf/compiler/code_generator.h>
 #include <google/protobuf/compiler/java/java_helpers.h>
 #include <google/protobuf/stubs/substitute.h>
