+++ src/google/protobuf/compiler/objectivec/objectivec_helpers.cc
@@ -41,6 +41,7 @@
 #include <unordered_set>
 #include <vector>
 
+#include <google/protobuf/compiler/code_generator.h>
 #include <google/protobuf/compiler/objectivec/objectivec_helpers.h>
 #include <google/protobuf/compiler/objectivec/objectivec_nsobject_methods.h>
 #include <google/protobuf/descriptor.pb.h>
@@ -360,14 +361,6 @@ std::string GetEnumNameForFlagType(const FlagType flag_type) {
 // Escape C++ trigraphs by escaping question marks to \?
 std::string EscapeTrigraphs(const std::string& to_escape) {
   return StringReplace(to_escape, "?", "\\?", true);
-}
-
-std::string StripProto(const std::string& filename) {
-  if (HasSuffixString(filename, ".protodevel")) {
-    return StripSuffixString(filename, ".protodevel");
-  } else {
-    return StripSuffixString(filename, ".proto");
-  }
 }
 
 void TrimWhitespace(StringPiece* input) {
